move getResourcesFromBindings to pipelineRun
